+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 * chore(databases): refactor constants
 * build(deps): update `github.com/golang-jwt/jwt` from v4 to v5
-* refactor: replace `errgo` with `github.com/Scalingo/go-utils/errors/v3`
+* refactor: replace `errgo` with `github.com/Scalingo/go-utils/errors/v3` (breaking change)
+* feat(logs): `LogsURL` returns a parsed structure (breaking change)
 
 ## 9.2.0